backend_DRF/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
django-cors-headers==4.2.0